refactoring : created a models package and added Order model in it.
