+++ main.py
@@ -1,20 +1,11 @@
-from typing import List, Optional
 from fastapi import FastAPI, Request, Response
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
-from pydantic import BaseModel
+from models import Order
 
 app = FastAPI()
 app.mount("/static", StaticFiles(directory="static"), name="static")
 templates = Jinja2Templates(directory="templates")
-
-
-class Order(BaseModel):
-    q: str
-    cuisineType: Optional[List[str]] = []
-    mealType: Optional[List[str]] = []
-    diet: Optional[List[str]] = []
-    health: Optional[List[str]] = []
 
 
 @app.get("/")
